Add slots for all equipment
There are a total of 14 accessory slots and 2 weapon slots in endgame
currently. The accessory slots are ordered 2-4-4-4, and added in a zig
zag manner, starting from the top right, and ending in the bottom left.
The second weapon slot is below the first weapon slot.

This commit adds the definition for all of them.